participant-count/requirements.txt
daily-python~=0.5.0
polling2~=0.5.0
requests~=2.31.0
